+++ leeedohyun/src/test/java/inflearn/mini/worktimehistory/service/WorkTimeHistoryServiceTest.java
@@ -19,8 +19,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import inflearn.mini.employee.controller.EmployeeWorkHistoryRequest;
-import inflearn.mini.employee.controller.EmployeeWorkHistoryResponse;
+import inflearn.mini.employee.dto.request.EmployeeWorkHistoryRequest;
+import inflearn.mini.employee.dto.response.DateWorkMinutes;
+import inflearn.mini.employee.dto.response.EmployeeWorkHistoryResponse;
 import inflearn.mini.employee.domain.Employee;
 import inflearn.mini.employee.exception.AbsentEmployeeException;
 import inflearn.mini.employee.exception.AlreadyAtWorkException;
